+++ example/shared/src/screens/index.ts
@@ -1,4 +1,7 @@
 export * from './MapScreen';
 export * from './ModalScreen';
+export * from './NotificationsSheetContent';
+export * from './ProfileSheetContent';
+export * from './SettingsSheetContent';
 export * from './StandardScreen';
 export * from './TestScreen';